+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/CachePoolClearerPass.php
@@ -11,31 +11,15 @@
 
 namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\Cache\DependencyInjection\CachePoolClearerPass as BaseCachePoolClearerPass;
+
+@trigger_error(sprintf('The "%s" class is deprecated since Symfony 4.2, use "%s" instead.', CachePoolClearerPass::class, BaseCachePoolClearerPass::class), E_USER_DEPRECATED);
 
 /**
  * @author Nicolas Grekas <p@tchwork.com>
+ *
+ * @deprecated since version 4.2, use Symfony\Component\Cache\DependencyInjection\CachePoolClearerPass instead.
  */
-final class CachePoolClearerPass implements CompilerPassInterface
+class CachePoolClearerPass extends BaseCachePoolClearerPass
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function process(ContainerBuilder $container)
-    {
-        $container->getParameterBag()->remove('cache.prefix.seed');
-
-        foreach ($container->findTaggedServiceIds('cache.pool.clearer') as $id => $attr) {
-            $clearer = $container->getDefinition($id);
-            $pools = array();
-            foreach ($clearer->getArgument(0) as $name => $ref) {
-                if ($container->hasDefinition($ref)) {
-                    $pools[$name] = new Reference($ref);
-                }
-            }
-            $clearer->replaceArgument(0, $pools);
-        }
-    }
 }